Solution to Find the Length of a String

+++ Find the Length of a String@@676c039d80e1714600f67643/code.c
@@ -3,7 +3,7 @@
 int main(){
     char str[100];
     scanf("%s",&str);
-    double length=strlen(str);
-    printf("%.0lf",length);
+    int length=strlen(str);
+    printf("%d",length);
     return 0;
 }
